Adding a pattern I've seen in the wild, calling it Listener.
This is something I have seen used in some cases where there is a thing
that wants to allow code written by others to be able to interact with
the client when something happens based on event types. In this way, one
can register a function to a given event handler so that, when the event
happens, the client calls the function, usually with any event-specific
information being passed to the function so that it is able to make more
informed decisions.

I wanted to document this because I wanted to get other people's take on
this and whether it's a true pattern, called something else and I am
just not aware, or what the deal is.